+++ WasmNet/Opcodes/NumericOpcodes/F32/F32CeilOpcode.cs
@@ -7,7 +7,7 @@ namespace WasmNet.Opcodes {
             return visitor.Visit(this, arg);
         }
 
-        protected override double Execute(float arg) => Math.Ceiling(arg);
+        protected override float Execute(float arg) => (float)Math.Ceiling(arg);
 
         public override string ToString() => "f32.ceil";
