+++ WasmNet/Opcodes/NumericOpcodes/F32/F32MinOpcode.cs
@@ -7,7 +7,7 @@ namespace WasmNet.Opcodes {
             return visitor.Visit(this, arg);
         }
 
-        protected override double Execute(float left, float right) => Math.Min(left, right);
+        protected override float Execute(float left, float right) => Math.Min(left, right);
 
         public override string ToString() => "f32.min";
